requirements.txt
bittensor
torch
openai
loguru
apify-client
elasticsearch
python-dotenv
nltk